Make 'xmlfile' optional (rt#42069)
TAP::Harness->new does not require and xmlfile parameter, but
TAP::Harness::JUnit->new requires it. This makes it impossible to use
TAP::Harness::JUnit as a drop-in replacement for TAP::Harness. This
functionality is necessary, for example, if one would like to use
TAP::Harness::JUnit as the harness used by prove. The supplied patch
defaults the xmlfile argument to "junit_output.xml" if the use does not
supply it, and issues a warning. A test for this behavior has been
added.