Do not call /bin/cat to get TAP output into parser (rt#41344)
Bug report:

System Environment:
Cygwin in c:\cygwin
PATH contains c:\cygwin\bin, but NOT c:\cygwin

Execution of $jharness->runtests(@tests); leads to an xml file which
is not filled with test result data, and the following cmd line msg:
"The system cannot find the path specified."

Debugging shows that the reason for this is a hardcoded path in
JUnit.pm, line number 95. If I replace '/bin/cat' by 'cat', the bug is
fixed.